Eccomerce/backend/requirements.txt
django==4.2
djangorestframework==3.15.2
django-extensions
djangorestframework
django-filter